mkrootfs: Check active QEMU binfmt interpreter on systemd-nspawn backend
This was previously only done on chroot backend as update-binfmts
provided a command to check this, but as systemd doesn't seem to have
such "niceties" just query this from procfs everywhere.

We still want to check for update-binfmts binary in deps however as it's
a simple to check if binfmt-support package is actually installed.